gdb: Introduce new setting to filter out shadowed variables.
Add a new setting 'print shadowed on|off' to control the printing of
shadowed variables in 'info locals'.  Add a new option '-shadowed
on|off' to override this setting for the invocation of the command.

When shadowed variables are present, each variable involved in a
shadowing relationship is annotated with its declaration location
'<file:line>'; the ones hidden by an inner declaration are additionally
marked ', shadowed'.

Given this code, stopped at line 10:

   1 int num = 1;
   2 int
   3 main ()
   4 {
   5   const char *str = "main";
   6   int num = 3;
   7   {
   8     const char *str = "nested";
   9     int num = 5;
  10     num = 0; //  break here
  11   }
  12   return num;
  13 }

By default, 'info locals' prints shadowed variables:

  (gdb) info locals
  str = 0x555555556009 "nested"  <main.c:8>
  num = 5  <main.c:9>
  str = 0x555555556004 "main"  <main.c:5, shadowed>
  num = 3  <main.c:6, shadowed>
  Use "set print shadowed off" to hide shadowed variables.

The user may not want to print shadowed variables:

  (gdb) set print shadowed off
  (gdb) info locals
  str = 0x555555556009 "nested"
  num = 5
  Some shadowed variables were omitted, use "set print shadowed on" to include them.

Approved-By: Tom Tromey <tom@tromey.com>

+++ gdb/testsuite/gdb.base/var-shadowing.c
@@ -24,7 +24,7 @@ shadowing (void)
   a = 101;  /* bp for locals 1 */
   {
     unsigned int val2 = 3;		/* val2-d2 */
-    unsigned int val3 = 4;		/* val3-d1 */
+    double val3 = 4;			/* val3-d1 */
     a = 102;  /* bp for locals 2 */
     {
       unsigned int val1 = 5;		/* val1-d2 */
